Lab4
COP 1335C
Lab 4
25 points
Character Stack
File names: charStack.h, charStackMain.cpp
a) (20 points)
Modify the integer stack that we designed in class so that it works for a
stack of characters. Rename the class CharStack.
b) (5 points)
Once your class is defined, write a small test program (in a separate file)
to thoroughly test all of the member functions of your stack.